refactor: extract language server implementation in its own module
