+++ internal/cmd/lsp.go
@@ -2,6 +2,7 @@ package cmd
 
 import (
 	"github.com/formancehq/numscript/internal/lsp"
+	"github.com/formancehq/numscript/internal/lsp/language_server"
 
 	"github.com/spf13/cobra"
 )
@@ -12,8 +13,8 @@ var lspCmd = &cobra.Command{
 	Long:   "Run the lsp server. This command is usually meant to be used for editors integration.",
 	Hidden: true,
 	Run: func(cmd *cobra.Command, args []string) {
-		lsp.RunServer(lsp.ServerArgs[lsp.State]{
-			InitialState: lsp.InitialState(),
+		language_server.RunServer(language_server.ServerArgs[lsp.State]{
+			InitialState: lsp.InitialState,
 			Handler:      lsp.Handle,
 		})
 	},
